Add async execution where appropriate
