+++ api/src/test/scala/hmda/api/http/InstitutionHttpApiAsyncSpec.scala
@@ -1,0 +1,5 @@
+package hmda.api.http
+
+import org.scalatest.AsyncWordSpec
+
+trait InstitutionHttpApiAsyncSpec extends AsyncWordSpec with InstitutionHttpSpec
